+++ climaf/classes.py
@@ -14,6 +14,7 @@ from period    import init_period, cperiod
 from clogging  import clogger, dedent
 from netcdfbasics import fileHasVar, varsOfFile, timeLimits, model_id
 from decimal   import Decimal
+import climaf
 
 #: Dictionary of declared projects (type is cproject)
 cprojects=dict()
@@ -1368,6 +1369,203 @@ def browse_tree(cobj,func,results):
         return
 
 
+def check_time_consistency(any):
+    """
+    Check time consistency of first variable of a dataset or ensemble members:
+     - check if first data time interval is consistent with dataset frequency
+     - check if file data have a gap
+     - check if period covered by data files actually includes the whole of dataset period
+
+    Arg: a dataset or an ensemble
+     
+    Returns: True if period of data files included dataset period, False otherwise.
+
+    Examples:
+    
+    >>> # Dataset with monthly frequency  
+    >>> tas=ds(project='example', simulation='AMIPV6ALB2G', variable='tas',period='1980-1981')
+    >>> res1=check_time_consistency(tas)
+    >>>
+    >>> # Ensemble with monthly frequency  
+    >>> j0=ds(project='example',simulation='AMIPV6ALB2G', variable='tas', frequency='monthly', period='1980')
+    >>> j1=ds(project='example',simulation='AMIPV6ALB2G', variable='tas', frequency='monthly', period='1981')
+    >>> ens=cens({'1980':j0, '1981':j1})
+    >>> res2=check_time_consistency(ens)
+    
+    >>> # Define a new project for 'em' data with 3 hours frequency in particular     
+    >>> cproject('em_3h','root','group','realm','frequency',separator='|')
+    >>> path='/cnrm/aster/data1/simulations/${group}/${realm}/Regu/${frequency}/${simulation}/${variable}_??_YYYY.nc' 
+    >>> dataloc(project='em_3h', organization='generic', url=path)
+    
+    >>> # Dataset with 3h frequency for 'tas' variable (instant) 
+    >>> tas_3h=ds(project='em_3h',variable='tas',group='AR4',realm='Atmos',frequency='3Hourly', simulation='A1B',period='2050-2100')
+    >>> res3=check_time_consistency(tas_3h)
+     
+    >>> # Dataset with 3h frequency for 'pr' variable (time mean) 
+    >>> pr_3h=ds(project='em_3h',variable='pr',group='AR4',realm='Atmos',frequency='3Hourly', simulation='A1B',period='2050-2100')
+    >>> res4=check_time_consistency(pr_3h)   
+    
+    """
+    from anynetcdf import ncf
+    from datetime import datetime, timedelta
+    from netCDF4 import num2date
+    import numpy as np
+
+    # Returns the list of files which include the data for the dataset
+    # or for each member of the ensemble
+    if isinstance(any,cens):
+        rep=True
+        for memb in any:
+            #clogger.info('Member: %s'%memb)
+            rep=check_time_consistency(any[memb]) and rep
+        return rep
+    elif isinstance(any,cdataset):
+        files=any.baseFiles()
+    	if not files:
+	    #clogger.error('No file found for: %s'%any)
+            return(False)
+    else :
+        clogger.error("Cannot handle %s" %any)
+        return
+    #
+    if files:
+        filedate=[]
+        clogger.debug("List of selected files: %s"%files)
+
+        var=str.split(climaf.driver.varOf(any),',')[0] 
+        # Concatenate all data files
+        for filename in str.split(files,' '):
+            fileobj=ncf(filename)
+            #
+            if any.project in aliases and var in aliases[any.project]:
+                var=aliases[any.project][var][0]
+            #
+            dimname=''
+            for dim in fileobj.variables[var].dimensions:  
+                if 'time' in dim: dimname=dim
+            if not dimname: clogger.error('No time dimension for variable %s'%var)
+            time_obj=fileobj.variables[dimname]  
+            filedate=np.concatenate((filedate,num2date(time_obj.getValue(),\
+                                   units=time_obj.units,calendar=time_obj.calendar)))
+            
+        clogger.debug('Time data of selected files: %s'%filedate)
+
+        # Check if first data time interval is consistent with dataset frequency
+        if len(filedate) > 1 :
+            filedate_delta=(filedate[1]-filedate[0]).total_seconds()
+        else:
+            clogger.error('Time dimension is degenerated.')
+            return
+
+        if ( (any.frequency == 'monthly' or not any.frequency) and \
+             (filedate_delta > 31.*24.*3600 or filedate_delta <= 29.*24.*3600.) ) \
+           or ( any.frequency == 'yearly' and \
+                (filedate_delta > 366.*24.*3600. or filedate_delta < 365.*24.*3600.) ) \
+           or ( any.frequency == 'decadal' and \
+                (filedate_delta > 3653.*24.*3600. or filedate_delta < 3651.*24.*3600.) ):
+        
+            clogger.warning('First data time interval (= %.1f days) is not consistent with dataset frequency (i.e. %s)'\
+                            %(filedate_delta/(24.*3600.),any.frequency))
+
+        elif any.frequency == 'daily' and filedate_delta != 86400.:
+            clogger.warning('First data time interval (= %.2f hours) is not consistent with dataset frequency (i.e. %s)'\
+                            %(filedate_delta/3600.,any.frequency))
+
+        elif (any.frequency == '6h'or any.frequency == '3h' or any.frequency == '1h' \
+              or any.frequency == '3Hourly' or any.frequency == '6Hourly') \
+              and filedate_delta != float(any.frequency[0])*3600.:
+            clogger.warning('First data time interval (= %.2f hours) is different to dataset frequency (i.e. %.2f)'\
+                            %(filedate_delta/3600.,float(any.frequency[0])))
+
+        # Check if file data have a gap
+        i=0
+        cpt=0
+        while i < len(filedate)-2:
+            i+=1
+            if (filedate[i+1]-filedate[i]).total_seconds() != filedate_delta:
+                cpt+=1
+                if cpt < 5:
+                    if any.frequency == 'monthly' or not any.frequency or \
+                       any.frequency == 'yearly' or any.frequency == 'decadal':
+                        clogger.error('File data have a gap between indexes %i and %i: delta = %.0f days '\
+                                      %(i,i+1,(filedate[i+1]-filedate[i]).total_seconds()/(24.*3600.)) +\
+                                      'instead of %.0f days (<=> 1st data interval)'\
+                                      %(filedate_delta/(24.*3600.)))
+                    elif any.frequency == 'daily' or any.frequency == '6h'or \
+                         any.frequency == '3h' or any.frequency == '1h' or \
+                         any.frequency == '3Hourly' or any.frequency == '6Hourly':
+                        clogger.error('File data have a gap between indexes %i and %i: '%(i,i+1) +\
+                                      'delta = %.0f hours instead of %.0f hours (<=> 1st data interval)' \
+                                      %((filedate[i+1]-filedate[i]).total_seconds()/3600.,filedate_delta/3600.))
+        #
+        # Compute period covered by data files
+        if any.frequency == 'monthly' or not any.frequency:  
+            filedate[0]=filedate[0].replace(day=01)
+            if filedate[-1].month > 11 :
+                filedate[-1]=filedate[-1].replace(year=filedate[-1].year+1)
+                filedate[-1]=filedate[-1].replace(month=01)
+                filedate[-1]=filedate[-1].replace(day=01)
+            else:
+                filedate[-1]=filedate[-1].replace(month=filedate[-1].month+1)
+                filedate[-1]=filedate[-1].replace(day=01)
+
+        elif any.frequency == 'daily':
+            filedate[0]=filedate[0].replace(hour=00)
+            filedate[-1]=filedate[-1].replace(hour=00)
+            filedate[-1]=filedate[-1] + timedelta(days=1)
+              
+        elif any.frequency == '6h'or any.frequency == '3h' or any.frequency == '1h' \
+                 or any.frequency == '3Hourly' or any.frequency == '6Hourly':
+           
+            if 'cell_methods' in fileobj.variables[var].__dict__ : # time mean 
+
+                regex=re.compile('.*time *: *mean *\(? *interval *: *([0-9]+.?[0-9]+?) ([a-zA-Z]+) *\)')
+                cell_meth_att = regex.search(fileobj.variables[var].cell_methods)
+                if cell_meth_att:
+                    if cell_meth_att.group(2) == 'hours': freq=float(cell_meth_att.group(1))
+                    elif cell_meth_att.group(2) == 'minutes': freq=float(cell_meth_att.group(1))/60.
+                else: # 'cell_methods' attribute defined with the value 'time: mean'
+                    freq=filedate_delta/3600.
+
+                filedate[0] = filedate[0] - timedelta( minutes=(freq/2.)*60 + \
+                                ((filedate[0].hour*60 + filedate[0].minute)-(freq/2.)*60)%(freq*60) )
+                filedate[-1] = filedate[-1] - timedelta( minutes=(freq/2.)*60 + \
+                                ((filedate[-1].hour*60 + filedate[-1].minute)-(freq/2.)*60)%(freq*60) - freq*60 ) 
+
+            else: # assume it is instant data
+                freq=filedate_delta/3600.
+                filedate[-1] = filedate[-1] - timedelta( minutes=(freq/2.)*60 + \
+                    ((filedate[-1].hour*60 + filedate[-1].minute)-(freq/2.)*60)%(freq*60) - 2*freq*60 )
+
+        elif any.frequency == 'yearly' or any.frequency == 'decadal':
+            filedate[0]=filedate[0].replace(month=01)
+            filedate[0]=filedate[0].replace(day=01)
+            filedate[-1]=filedate[-1].replace(month=01)
+            filedate[-1]=filedate[-1].replace(day=01)
+            filedate[-1]=filedate[-1] + timedelta(years=1)
+   
+        elif any.frequency == 'fx' or any.frequency == 'annual_cycle':
+            clogger.error('Check time consistency with a frequency equal to %s has no sense' %any.frequency)      
+
+        else:
+            clogger.error('Dataset frequency is non-standard: frequency = %s. ' %any.frequency +\
+                          'Normalized frequency values are: decadal, yearly, monthly, '+\
+                          'daily, 6h, 3h, fx and annual_cycle')
+        #
+        # Check period of datafiles vs dataset period
+        clogger.debug('Period covered by selected files: %s'%filedate)
+        file_period=cperiod(start=filedate[0],end=filedate[-1])
+        #
+        if file_period.includes(any.period):
+            clogger.info("Time data in datafiles (i.e. %s) includes time data of " %file_period +\
+                         "dataset (i.e. %s) => dataset are consistent." %any.period)
+            return(True)
+        else:
+            clogger.info("Time data in datafiles (i.e. %s) don't include time data of " %file_period +\
+                         "dataset (i.e. %s) => dataset are not consistent." %any.period)
+            return(False)
+
+        
 class Climaf_Classes_Error(Exception):
     def __init__(self, valeur):
         self.valeur = valeur